backend/requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
python-multipart>=0.0.6
pydantic>=2.0.0
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
alembic>=1.10.4

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# AI Integration
openai>=0.27.0

# Voice Processing
SpeechRecognition>=3.10.0
pydub>=0.25.1

# Payment Integration
stripe>=5.0.0

# Testing
pytest>=7.3.1
httpx>=0.24.0

# Utilities
tenacity>=8.2.2
pydantic-settings>=2.0.2
